system-design/kafka-example/requirements.txt
kafka-python==2.0.2
asyncio==3.4.3
colorama==0.4.6 